refactor: separate bridge monitoring types and status query
